+++ cppad/local/optimize/record_pv.hpp
@@ -90,7 +90,7 @@ struct_size_pair record_pv(
 	//
 	struct_size_pair ret;
 	ret.i_op  = rec->num_op_rec();
-	ret.i_var = rec->PutOp(op);
+	ret.i_var = size_t(rec->PutOp(op));
 	CPPAD_ASSERT_UNKNOWN( 0 < new_arg[1] && size_t(new_arg[1]) < ret.i_var );
 	return ret;
 }
